[hotfix][network] Move private NoOpResultSubpartitionView class out of PartitionRequestQueueTest
NoOpResultSubpartitionView is defined as a private class in PartitionRequestQueueTest. It could be used widely
in other places future, so it is reasonable to refactor it as a separate public class.
